requirements.txt
# Frontend
streamlit

# Backend
fastapi
uvicorn

# Agent & LLM
langchain
langchain-google-genai
google-generativeai

# Google Calendar Integration
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv

# General Utilities
requests
